Introduce tapir-cats-vertx-server module
